Use CoreAnimation fences instead of synchronous IPC to synchronize resize https://bugs.webkit.org/show_bug.cgi?id=146294 <rdar://problem/21090193>
Reviewed by Anders Carlsson.

* UIProcess/mac/TiledCoreAnimationDrawingAreaProxy.mm:
(WebKit::TiledCoreAnimationDrawingAreaProxy::createFenceForGeometryUpdate):
Tiny followup; don't touch Connection if the connection isn't valid.


Canonical link: https://commits.webkit.org/164930@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186693 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/mac/TiledCoreAnimationDrawingAreaProxy.mm
@@ -153,6 +153,9 @@ void TiledCoreAnimationDrawingAreaProxy::willSendUpdateGeometry()
 MachSendRight TiledCoreAnimationDrawingAreaProxy::createFenceForGeometryUpdate()
 {
 #if HAVE(COREANIMATION_FENCES)
+    if (!m_webPageProxy.isValid())
+        return MachSendRight();
+
     RetainPtr<CAContext> rootLayerContext = [asLayer(m_webPageProxy.acceleratedCompositingRootLayer()) context];
     if (!rootLayerContext)
         return MachSendRight();
@@ -173,6 +176,8 @@ MachSendRight TiledCoreAnimationDrawingAreaProxy::createFenceForGeometryUpdate()
     });
     RefPtr<WebPageProxy> retainedPage = &m_webPageProxy;
     [CATransaction addCommitHandler:[callbackID, retainedPage] {
+        if (!retainedPage->isValid())
+            return;
         if (Connection* connection = retainedPage->process().connection())
             connection->uninstallIncomingSyncMessageCallback(callbackID);
     } forPhase:kCATransactionPhasePostCommit];
